refactor(generator): extract controls sheet helpers
